add support for in/out GPIO pins on ULX3S
The ULX3S has 2x28 IO pins that can be used either individually or as
differential pairs. Currently we're going to assume an individual setup (you can
configure their behaviour in the .lpf file).

The convention is to see them as gp[27] (GPIO positive) and gn[27] (GPIO
negative) arrays; this is how they're numbered on the board itself. Note that
positive and negative moniker are irrelevant in the current configuration as the
ports aren't set up as differential pairs. To work with this convention, we
split out the GPIO address space into 0x00XX and 0x02XX for setting in/output
values of gp and gn pins respectively, and use 0x01XX and 0x3XX for setting the
direction of the gn and gp pins.